requirements.txt
pycryptodome==3.16.0
python-dotenv==0.21.0
